agenticragyoutube/requirements.txt
streamlit
langchain-core
langchain-community
langchain-openai
langchain-classic
langgraph
faiss-cpu
python-dotenv
youtube-transcript-api
tiktoken